pyproject.toml
[project]
name = "reporium-events"
version = "1.0.0"
description = "Event publishing client for the Reporium platform — GCP Pub/Sub"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-pubsub>=2.20",
    "google-cloud-firestore>=2.16",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
